feat: add batch handling for the quick indexer

+++ migrations/20250107071623_index_metadata.up.sql
@@ -1,6 +1,7 @@
+-- Add up migration script here
 CREATE TABLE
     IF NOT EXISTS index_metadata (
-        id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+        id INT8 GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
         current_latest_block_number BIGINT NOT NULL,
         indexing_starting_block_number BIGINT NOT NULL,
         is_backfilling BOOLEAN DEFAULT TRUE,
